requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-autoslug==1.9.9
django-debug-toolbar==4.3.0
django-filter==24.2
django-sortedm2m==3.1.1
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
gunicorn==22.0.0
idna==3.7
oauthlib==3.2.2
packaging==24.0
psycopg==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==69.5.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
wheel==0.43.0
